Implemented IShellDetails in order to display properties of the host connections when the window is in Details view on all Windows versions since 95 (tested on XP and 98 so far).  The text for these details is displayed but the default column widths seem to vary widly from what has been set programatically and differ also between OSes (XP and 98).  It is unclear how to solve this.
In an attempt to get detail to appear as subitems in an XP tiles view, I implemented IShellFolder2 which supports advanced details using SHCOLUMNIDs/PROPERTYKEYs.  This has been done using custom Property IDs (PIDs) and it hasn't caused any problems using Swish on a Windows 98 machine (Windows 98 does not support IShellFolder2).

The next step is to make the legacy IShellDetails functions call the property-based IShellFolder2 functions using MapColumnToSCID to convert between them.  This may have to be reverted if it breaks support for Windows 98.