IPersistFolder2 implemented causing XP WebView Details 'exapando' to appear.  Why this is so is not at all clear.  This touches on the issue but in reference to IPersistFolder3: http://www.ureader.com/message/695263.aspx
Most columns/details are now chosen from built-in System core properties.  The two that aren't, Username and Port are now custom PKEYs.

Added support for IQueryAssociation from GetUIObjectOf.  In conjunction with the TileInfo registry key, this causes the properties listed in the key to be shown as subitem details when in Tile view.  At the moment this doesn't seem to work if the registry key uses named properties so we are using the {fmtid}pid arrangement to show the Host and Path details.  AssocCreate is used to direct any search for the TileInfo key to the SwishFolder CLSID.

Extensive refactoring of GetDetailsEx and MapColumnIdToSCID allowing use of system PKEYs.
GetDetailsOf now delegates any real work to GetDetailsEx by using MapColumnIdToSCID and marshalling the result from the returned VARIANT into the wchar string expected.

Tested on Windows XP and Windows 98. A script error appears if one of the host connection items is selected in the Swish window under Windows 98.  This hadn't been tested before and is unlikely to be a regression.