Added RemoteFolder class to represent remote folder listings.  Currently, this folder displays a single file in a dummy file listing. The folder is displayed by double-clicking (or opening via the Explorer tree or context menu) on a host connection object in the HostFolder.
The items in the folder (currently one dummy PIDL) are enumerated by the RemoteEnumIDList class.  For the moment this class simulates connection to a remote host but eventually such duties should be given to a separate class with RemoteEnumIDList responsible for enumeration of PIDLs alone.

The display and execution of the host object context menu is handled by HostContextMenu.  The creation of a context menu was necessary to support double-clicking: the DefView ShellView reacts to a double-click by invoking the objects default context menu command, unlike the Explorer tree which calls BindToObject to obtain the new ShellFolder object.

RemoteFolder has some redundant initialisation code (_init) and the correct initialiser (Initialize) is not being called for any sub-folders.  Dealing with this is the next task.