Extended Rename() to cope if the file already exists.  This was tricky because SFTP servers < v4 (i.e. most) don't allow an overwrite in this case.  We get around this by first trying the v5 method and if that fails trying a non-atomic delete-rename combination.  This has the potential for race conditions but I think the benefits outweigh the risk.  If this proves a problem I'll remove it and just make it report the failure.
Also, more than one operation could not be performed down a single connection.  I have hacked in fix for this (m_fConnected) but eventually it should be handled by the connection pool that I want to make.

Was asserting if EnumObjects() UI was forbidden (NULL HWND passed).  Now we just throw an exception which gets caught as a simple failure.

Removed prompt strings from OnConfirmOverwrite methods.  This should really be up to the frontend not the SftpProvider.

ComCatch was not catching AtlExceptions.  Fixed.

Added unit tests for rename functionality.

Extended MockSftpConsumer so that any behaviours which are not explicitly set throw and exception.  This causes tests to fail if an unexpected callback is triggered.