Fixed const-related problems with CPidl class.  It was using 'const PidlType' which was actually a const pointer to a non-const PIDL.  It should have been using const *IdListType.  This has been corrected.
Added an operator which returns the PIDL as it's supertype.  This is to allow assignments like 'CRemotePidl r = CChildPidl c' to work in just the same way as 'PIDLIST_RELATIVE r = PITEMID_CHILD' would.

Fixed throw() indications.

Added an Join() method to join another PIDL to the CPidl.  In the case of joining to a CChildPidl this should really return a CRelativePidl but I'm not sure how to do this with templates.  this is left for the future.

Added a GetNext() method to return a const pointer to the next SHITEMID in the member PIDL.  This doesn't make sense to use on a CChildPidl.

Added unit tests to cover the new cases.