Reordered parameters of CHostPidl and CRemotePidl creator constructors.  This removes an ambiguity with the concatenation contructors.
Added GetFullPath() method to CHostPidl which returns the absolute path of a file on the remote filesystem, e.g., "/tmp/dir1/dir2/filename.ext".  Unfortunately, the gives HostPidl.h a dependency on RemotePidl.h.
Added GetFilePath() method to CRemotePidl which returns the relative path of the file represented by the pidl.  This will simply be the filename for a child pidl.  E.g. a pidl with three items in it could give: "dir1/dir2/filename.ext".