Added unit tests for CPool.
The tests were added to a new Boost.Test-based test-shell_folder project and the old tests are now in test-shell_folder-legacy.  The new tests use a fixture that fires up a local instance of OpenSSH (sshd) and uses that for 'network' testing.  This should be much faster than the old way which relied on an actual remote host being available.