Reduced dependencies between projects.
In an effort to prevent the continual rebuilding/rerunning of unit tests in unrelated projects I have removed any dependencies where one project depends on another whose output is a COM component.  These components are not linked with the dependent projects so are not required as references.  However, as some of the COM component projects generate files that the others use in their compilation, it may now take two or more runs to completely build the solution when using a pristine source tree.