Fixed memory leaks by shutting down the sessions properly.
It's a bit ambiguous in the libssh2 documentation, but we have to call libssh2_sftp_shutdown() if we have set up and SFTP session in order to reclaim its resources properly.  A call to libssh2_session_disconnect(), while not required, is good etiquette and prevents the server reporting "Read error from remote host 127.0.0.1: Connection reset by peer".