Refactored the folders so that GetUIObjectOf and CreateViewObject requests are routed and delegated properly.
Now we delegate all GetUIObjectOf requests for folders to the folder instance's CreateViewObject method rather than trying to answer some queries on its behalf.  Also, GetUIObject calls without any PIDLs given are for the current folder rather than a child item and are routed to the same code as calls to CreateViewObject on the current folder instance.