Replaced addressof override in the StorageMedium.
Overriding an address operator is bad practice and leads to all sorts of nastyness such as preventing STL collections containing the type.  I have replaced it with an explicit out() method that renders a pointer to the wrapped data member.  This is in the spirit of the Comet library classes.