Conditional PCH blocks.
Added conditional block to the common precompiled header so that projects that don't use Boost don't incur the build time penalty of precompiling the Boost headers.

Also, added a line to the test-common_boost project that changes the permissions on the host key file using 'cacls' so that OpenSSH doesn't complain.