Fixed bind_to_handler_object() to work with NULL or empty PIDLs.
When a NULL or empty PIDL is passed to bind_to_handler_object() (or SHBindToObject()) the callers intention is most likely to request the handler for the Desktop folder.  However, as this function was implemented in terms of the desktop folder already, passing NULL or an empty PIDL to its BindToObject() method caused it to fail.  Now we catch this case and use QueryInterface to return the handler rather than BindToObject().