Add progress display to DropTarget.cpp
To displaying progress we need to how many items are to be copied before we start to copy them.  This meant the algorithm had to be changed from one that copied as it recursively iterated down the hierarchy to one that, first, recursively builds a list of items to copy and then copies them.