Changed SftpStream error HRESULTs that map to a message with %-signs.
When creating an SftpStream a number of errors can occur such as file-not-found.  These are mapped to HRESULTs which are returned to Explorer which displays the associated message text (as looked up by FormatMessage()) in a message box.  However, if the message contains substitutable sections (i.e., printf-style '%' markers) Explorer won't display the error and the operation fails silently without informing the user that anything is amiss.

This change removes these kinds of error codes from the ones that SftpStream can return.