User Boost.Asio to resolve addresses instead of raw Winsock.
This is for compatibility with Windows 98/Me which don't include the getaddrinfo functions.  These used to be provided by the SDK instead but since SDK 6.1 this is #ifdeffed only for versions of windows greater than 2k.

We define the BOOST_ASIO_DISABLE_IOCP macro as the CreateIoCompletionPort function is present but doesn't seem to work on Windows 98.  This macro selects an alternative implementation.