Add type-checking.
Whenever a basic_pidl is created from a raw PIDL, either by construction or using attach(), the PIDL is checked to make sure that it matches a type constraint defined by the traits class.

In practice, this means that child PIDLs check that there is only one item in the incoming raw PIDL and throw an exception otherwise.