Allow DropTarget to cancel a transfer mid-file.
When transferring a large file from the local machine to the remote server there was no way to cancel part-way through; cancelling only worked between files.  This was a problem if the files were large.

This fix changes CDropTarget so it copies files in 32K chunks and checks for cancellation between each one.