Stop storing consumer pointer in provider instance.
Storing an ISftpConsumer instance in the provider via SwitchConsumer is inherently dangerous.  The provider can be used from many apartments and if the apartment the consumer was created in has ended when a user in another use tries to SwitchConsumer, CProvider tries to Release the old, dead consumer causing the program to hang.

ISftpConsumer should eventually be replaced by a different mechanism (a combination of IErrorInfo for reporting error messages and custom interfaces in those few situations that need two-way communication) but a rough fix for the moment is to pass a consumer to every ISftpProvider call that's valid for the duration of that call only.  This is a messy fix that touches many many files.