Fix GUI freeze when pressing Alt+Space in a dialogue.
Alt+Space invokes the system menu which is created there and then.  Previously, our window creation hook was still installed and would capture this.  While this should work, for some reason the menu has data (not ours) in lpCreateParams.  When we try and cast this to a window_impl pointer we die.  

This fix reduced the amount of time we hook window creation to an absolute minimum.  We install the hook just before calling DialogBoxIndirectParam and remove it when we receive the WM_INITDIALOG message.  This is *before* DialogBoxIndirectParam even returns!  Any windows created after this point won't be our dialog controls.

Still, it's possible that the dialog could itself create another dialog so we might need to make this even safer somehow.

Fixes #45.