Fix context menus in Windows 7.
CDefFolderMenu_Create2 is much more sensitive in Windows 7 than in previous versions.  Firstly, we have to get the error codes exactly right.  Secondly, the order of the passed-in registry keys matters.  We have to pass HKCR\Folder before HKCR\Directory for the 'Open' verb to be chosen as the default.

Fixes #27.