Webtasks conflicting with ExplorerCommands.
On Windows 7 (and presumably Vista) Explorer takes a look at the (undocumented) XP-style webtasks as well as the (documented) ExplorerCommands.  Something about our webtasks upsets Explorer so much that it won't even display the ExplorerCommands any more.

It probably looks at a table of supported webtasks to try and map them to ExplorerCommands and fails to find our command GUIDs in it as this undocumented feature wasn't meant to be used outside Microsoft.

The solution is to sniff the Windows version and pretend that we don't have webtasks when running on Vista or later.