Added missing TaskDialog enums.
We defined some of the TaskDialog-related defines/enums/structs but this caused problems for other code that tried to do the same.  If other code decides whether it need to define one of the enums we omitted based on the presence of a #define we provided then it would end up with an incomplete set of TaskDialog enums.