Fix task_dialog throwing an exception when no button specified.
If no buttons were specified for the task_dialog, the TaskDialog implementation added one anyway with a caption OK and code IDOK.  When the user clicked this it would try to find the callback in m_callbacks and end up calling a default-constructed boost::function.  This function would be empty and therefore threw an exception.

Now we create a noop callback with IDOK in m_callbacks if no buttons were explicitly added.

See #83.