Support ampersand (&) accelerators in expando text.
The only way I could find to do this was to add a hidden button which gets clicked when the user presses the adjacent static label's mnemonic.  This is necessary because the expando button itself is a static icon control.  These don't get clicked when their adjacent mnemonic is triggered; Windows just gives them focus.  That's not enough for us.

I've made the hidden window invisible by setting its rectangle to 0 x 0.  This is because properly hiding it with ShowWindow also disables it and it no longer responds as though it had been clicked.  There may be a better way to do this.