Fix layout in case where there are no conventional buttons.
If, for instance, all the buttons on the TaskDialog were command links, the layout was incorrect: a grey area appeared at the bottom of the dialogue.  This box was made from the padding which would have appeared round the non-existent conventional buttons.

This change fixes the layout by only adding the padding if something (buttons, expando, footer, etc.) appears below the white/grey fold.