Fix problem with Explorer windows losing and regaining focus (flashing).
This was a hard one to crack!  Turns out The ClassMoniker created by MkParseDisplayName (called inside CoGetObject) created a dialogue window (FFS!) saying "Installing COM Object" and owned by the desktop window.  it never showed the dialogue so the only symptom was this loss of focus.

The solution was to implement our own version of CoGetObject that affords us access to the IBindCtx.  This way we could set the IBindStatusCallback to a custom one which forbids UI.

This problem, filed in #80, became a blocker for #38 as it prevented focus landing on the newly created folder so the user could rename it.

Fixes #80.
See #38.