Display errors to user when drag-and-drop fails.
This is a complete hack but Explorer doesn't leave us much choice.  We have to show UI from the *data object* (madness!) because Explorer doesn't give us any better way (or if there is it isn't documented) and we have to do it without an HWND so there is no way let us know we should suppress UI in a particular case.