Prevent stack overflow if adapted command calls own methods.
This changes the adapter to aggregate the adapted command rather than inheriting from it because the latter can lead to a stack overflow if the superclass calls the adapted method in its implementation.  In other words, the adapter method would delegate to a method of the inner which, via virtual method dispatch, ends up delegating back the the adapter method.