Fix crash in Windows 8 drag-and-drop.
Windows 8 file-system IStreams have a bug where a call to Stat() doesn't set the pwcsName field even though STATFLAG_DEFAULT is passed.  This NULL pointer that is never supposed to be null caused Explorer to crash when we returned it as a string.  We never even used the string!

This change stops the code requesting the name so it just uses the file size, avoinding the problem altogether.

Fixes #221.