A try at fixing the blocking UI when doing a synchronous transfer.
Because we are no longer doing the transfer in a different COM apartment, which would pump messages during the call, the UI blocks on the drop.  That includes not showing the progress dialog.

This change attempted to create and update the progress dialog on a new thread.  It doesn't work because, I think, it breaks the Windows windowing rules by having an owned window created (and pumped) in a different thread.  The pumping has to happen in the main thread.