Pump UI messages on synchronous drop.
Because we are no longer doing the transfer in a different COM apartment, which would pump messages during the call, the UI blocks on the drop.  That includes not showing the progress dialog.

Therefore, we pump outstanding messages every time there is an update.  I don't think this it the right solution, but we can't run the progress dialog in a different thread as that breaks the windows rules.

The UI is still not wonderfully responsive because it can only update a little each time the progress is updated.  We may be able to do better once we use libssh2's non-blocking API as then we can pump messages more frequently.

Closes #182.