Only hide extensions of known file types.
The previous change was hiding the extension of all file, if the user had "Hide extensions for known file types" set.

This change looks up each file's type in the registry to decide whether to show the extension based on the presence of the AlwaysShowExt value.  The "Unknown" file type includes that value, so this change fixes the problem and also takes care of other file types that should always show their extension regardless of user setting.

Closes #90.