Fixed Launch Agent menu command state.
Unlike the Launch Agent command in the rest of the Explorer window, the Launch Agent command in Explorer Tools menu wasn't being disabled if the agent was already running.

This change fixes that problem.  The cause was that 'hidden' and 'disabled' were not mutually-exclusive and the LaunchAgent command set one but not the other.  The menu code then checked the *other* state.  This change makes the two states mutually exclusive to prevent that situation.