Added command to close running sessions.
The command removes the session from the pool.  However, any existing reference to the session will not be affected so running tasks will continue and the session will only be closed once they are complete.

We should provide a way for the user to truly log out and kill even those sessions that are in use.  This also relates to a bug where the session is in the pool but has not been authenticated (probably because authentication failed).  At the moment, these sessions appear to the user to be connected.

Addresses #85.