Refactored known-host write-line use `ssh::detail::libssh2::knownhost` namespace.
Moved it to the namespace as we are trying to move all calls to the libssh2 library there.  This also meant changing the error reporting to use the `system_error`s used in that namespace.  Made necessary changes to `ssh_error_category` to distinguish insufficient buffer space from other errors in a portable manner.