Fix session re-animation.
When sessions die we were trying to insert a freshly-connected one into the pool.  However, we were doing this by calling `insert()`, which doesn't replace existing items.  The result was that a new session would be created and authenticated and then destroyed, and then the old session was served up again.

This change fixes the code to use `replace()` instead of `insert()` when the session already exists.