Introduce command_site to manage UI ownership.
The command_site can take a fallback HWND, in addition to the OLE site,
to maximise the chance that we have an available UI owner.  This is an
abstraction of the UI-ownership available at the four different place a
command can me invoked: IContextMenu, IExplorerCommand, ISFVCallback
context menu, and WinXP task pane (IUICommand).